Adds nokogiri dependency.
A library change is causing newer HTML reports to return HTML with
IE9-specific tags. REXML cannot parse these tags (are they valid XML?),
so in order to be able to return ad hoc HTML reports, this dependency is
being added.

If this dependency proves to be a smooth transition, we may wish to
convert our parsing code to using it.