Removes addl XML parse step on report generation.
Code was needlessly parsing XML and HTML reports through REXML and
Nokogiri. It will be faster and less memory intensive to just pass the
string along, since #to_s was called on the resulting parse. If
we need to validate the XML for reports, that should be done in testing.