Add experimental method to adhoc scan a subset of assets
This behaves like the "Scan Now" dialog in the UI where you can select a
subset of a site's defined assets, a scan template, and a scan engine to
scan them with. These settings are not persisted to a site configuration
and allows multiple simultaneous scans on a single site.