When passed a filename, download a report in chunks
Rather than pulling an entire 6-8 GB report into a string in memory to
immediately write it out to a file, use the built-in feature of Net::HTTP to
retrieve the result in chunks and write them out as they are read from the
remote host. This prevents receiving `NoMemoryError` issues when downloading
huge reports.